Подчистил ошибки, сделал попап в каталоге по нажатию на иконки корзинок
