+++ .gitignore
@@ -9,3 +9,4 @@ css/*.css
 node_modules/
 npm-debug.*
 Thumbs.db
+build/
