+++ gulpfile.js
@@ -6,6 +6,7 @@ var plumber = require("gulp-plumber");
 var postcss = require("gulp-postcss");
 var autoprefixer = require("autoprefixer");
 var server = require("browser-sync").create();
+var minify = require("gulp-csso");
 
 gulp.task("style", function() {
   gulp.src("less/style.less")
@@ -15,6 +16,9 @@ gulp.task("style", function() {
       autoprefixer()
     ]))
     .pipe(gulp.dest("css"))
+    .pipe(minify())
+    .pipe(rename("style.min.css"))
+    .pipe(gulp.dest("css"));
     .pipe(server.stream());
 });
 
@@ -49,4 +53,27 @@ gulp.task("webp", function () {
   return gulp.src("img/**/*.{png,jpg}")
     .pipe(webp({quality: 90}))
     .pipe(gulp.dest("img"));
+});
+
+var imagemin = require("gulp-imagemin");
+
+gulp.task("images", function () {
+  return gulp.src("img/**/*.{png,jpg,svg}")
+    .pipe(imagemin([
+      imagemin.optipng({optimizationLevel: 3}),
+      imagemin.jpegtran({progressive: true}),
+      imagemin.svgo()
+    ]))
+    .pipe(gulp.dest("img"));
+});
+
+var posthtml = require("gulp-posthtml");
+var include = require("posthtml-include");
+
+gulp.task("html", function () {
+  return gulp.src("*.html")
+    .pipe(posthtml([
+      include()
+    ]))
+  .pipe(gulp.dest("."));
 });
