+++ less/blocks/contacts.less
@@ -186,7 +186,6 @@
 
   .contacts__map {
     position: absolute;
-    min-width: 0;
     width: 525px;
     height: 457px;
     left: 623px;
