+++ less/blocks/menu-toggle.less
@@ -15,6 +15,7 @@
 }
 
 .menu-toggle__icon {
+  display: block;
   position: relative;
   width: 20px;
   height: 2px;
