+++ less/blocks/production.less
@@ -55,14 +55,17 @@
   background: url("../img/icon-play.svg") no-repeat center;
   left: 78px;
   top: 36px;
+  transition-duration: 0.3s;
 
   &:hover {
     transform: scale(1.2);
+    transition-duration: 0.3s;
   }
 
   &:active {
     transform: scale(0.8);
     opacity: 0.3;
+    transition-duration: 0.3s;
   }
 }
