+++ less/blocks/user-menu.less
@@ -51,6 +51,16 @@
   fill: #63d1bb;
 }
 
+.user-menu__search:focus .user-menu__icon,
+.user-menu__search:hover .user-menu__icon {
+  fill: #63d1bb;
+}
+
+.user-menu__cart:focus .user-menu__icon,
+.user-menu__cart:hover .user-menu__icon {
+  fill: #63d1bb;
+}
+
 .user-menu__item:active .user-menu__icon {
   fill: #d0f1ea;
 }
