+++ package.json
@@ -8,9 +8,12 @@
     "autoprefixer": "7.1.x",
     "browser-sync": "2.18.x",
     "gulp": "^3.9.1",
+    "gulp-csso": "^3.0.0",
+    "gulp-imagemin": "^3.4.0",
     "gulp-less": "3.3.x",
     "gulp-plumber": "1.1.x",
     "gulp-postcss": "7.0.x",
+    "gulp-posthtml": "^3.0.0",
     "gulp-rename": "^1.2.2",
     "gulp-svgstore": "^6.1.0",
     "gulp-webp": "^2.3.0",
